# contracts/openapi/api.v1.yaml
# Generated from Solution.Api (OpenAPI v1). Do not edit by hand — run:
#   yarn --cwd web api:export
components:
  schemas:
    LoginUserRequest:
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
      type: object
    LoginUserResponse:
      properties:
        displayName:
          type: string
        email:
          type: string
        userId:
          format: uuid
          type: string
      required:
        - userId
        - displayName
        - email
      type: object
    ProblemDetails:
      properties:
        detail:
          type:
            - "null"
            - string
        instance:
          type:
            - "null"
            - string
        status:
          format: int32
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - "null"
            - integer
            - string
        title:
          type:
            - "null"
            - string
        type:
          type:
            - "null"
            - string
      type: object
    RegisterUserRequest:
      properties:
        displayName:
          type: string
        email:
          type: string
        password:
          type: string
      required:
        - displayName
        - email
        - password
      type: object
    UserRegistrationResponse:
      properties:
        createdAt:
          format: date-time
          type: string
        displayName:
          type: string
        email:
          type: string
        userId:
          format: uuid
          type: string
      required:
        - userId
        - displayName
        - email
        - createdAt
      type: object
info:
  title: Solution.Api | v1
  version: 1.0.0
openapi: 3.1.1
paths:
  /api/auth/login:
    post:
      operationId: LoginUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginUserRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginUserResponse"
          description: OK
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
          description: Bad Request
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
          description: Unauthorized
        "422":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
          description: Unprocessable Entity
        "500":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
          description: Internal Server Error
      tags:
        - Auth
  /api/auth/logout:
    post:
      operationId: LogoutUser
      responses:
        "204":
          description: No Content
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
          description: Unauthorized
      tags:
        - Auth
  /api/auth/me:
    get:
      operationId: GetCurrentUser
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginUserResponse"
          description: OK
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
          description: Unauthorized
      tags:
        - Auth
  /api/health:
    get:
      operationId: GetApiHealth
      responses:
        "200":
          description: OK
      tags:
        - Health
  /api/users:
    post:
      operationId: RegisterUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterUserRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserRegistrationResponse"
          description: Created
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
          description: Bad Request
        "422":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
          description: Unprocessable Entity
        "500":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
          description: Internal Server Error
      tags:
        - Users
tags:
  - name: Health
  - name: Auth
  - name: Users